xattr: exclude openbsd & solaris
