+++ fsimpl/localfs/system_unix.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"syscall"
 
-	"github.com/hugelgupf/p9/fsimpl/xattr"
 	"github.com/hugelgupf/p9/p9"
 	"golang.org/x/sys/unix"
 )
@@ -38,20 +37,4 @@ func (l *Local) lock(pid int, locktype p9.LockType, flags p9.LockFlags, start, l
 	}
 
 	return p9.LockStatusOK, nil
-}
-
-func (l *Local) SetXattr(attr string, data []byte, flags p9.XattrFlags) error {
-	return unix.Setxattr(l.path, attr, data, int(flags))
-}
-
-func (l *Local) ListXattrs() ([]string, error) {
-	return xattr.List(l.path)
-}
-
-func (l *Local) GetXattr(attr string) ([]byte, error) {
-	return xattr.Get(l.path, attr)
-}
-
-func (l *Local) RemoveXattr(attr string) error {
-	return unix.Removexattr(l.path, attr)
 }
